+++ contracts/nois-gateway/src/state.rs
@@ -1,6 +1,7 @@
 use cosmwasm_schema::cw_serde;
 use cosmwasm_std::{Addr, StdResult, Storage};
 use cw_storage_plus::{Deque, Item, Map};
+use nois_protocol::RequestBeaconOrigin;
 
 #[cw_serde]
 pub struct Config {
@@ -17,10 +18,7 @@ pub struct Job {
     pub source_id: String,
     // The channel the job came from and we have to send the response to
     pub channel: String,
-    // contract address on the app chain
-    pub sender: String,
-    /// A job ID assigned by the caller
-    pub job_id: String,
+    pub origin: RequestBeaconOrigin,
 }
 
 #[inline]