+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - drand: Only pay out rewards for bots in the right group ([#147]).
 - drand: Store height and tx_index of submission to allow finding transaction for a submission ([#153]).
 - proxy: Make callback gas limit configurable and reduce value to 500k in tests.
+- Pull out `RequestBeaconOrigin` struct which belongs to the proxy-dapp communication.
 
 [#147]: https://github.com/noislabs/nois-contracts/pull/147
 [#153]: https://github.com/noislabs/nois-contracts/issues/153